LayoutTests/platform/glib/imported/w3c/web-platform-tests/cookie-store/serviceworker_cookieStore_cross_origin.https.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: ReferenceError: Can't find variable: cookieStore

Harness Error (TIMEOUT), message = null

TIMEOUT cookieStore.get() in ServiceWorker reads cookie set in cross-origin frame Test timed out
